requirements.txt
opencv-python==4.9.0.80
streamlit==1.34.0
pillow==10.3.0